refactor: formatDeadline → format.ts 추출 + 테스트 3개 추가 (98→101개)
